% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_head.R
\name{access_head}
\alias{access_head}
\title{Convert YAML to an accessible header}
\usage{
access_head(rmd_path = NULL, lan = NULL, inplace = FALSE)
}
\arguments{
\item{rmd_path}{Path to the Rmd that requires accessible header metadata. Rmd
must be output type html.}

\item{lan}{Identify the language of text content.}

\item{inplace}{When set to FALSE (the default) writes to new file
(accessrmd_<rmd_path>). If TRUE, writes in place.}
}
\value{
Adjust the Rmd YAML provided to \code{rmd_path}, improving its
accessibility for screen readers. Only works with html output.
}
\description{
Reads an Rmd file, converting the YAML header to a format that is screen
reader friendly.
}
