% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_header.R
\name{assemble_header}
\alias{assemble_header}
\title{Assemble HTML header output from text provided.}
\usage{
assemble_header(title, subtitle = NULL, auth, doc_date, enc)
}
\arguments{
\item{title}{Text string to use as document title and h1 heading.}

\item{subtitle}{Text string to use as subtitle heading.}

\item{auth}{Text string or inline code to include as author heading.}

\item{doc_date}{Text string or inline code to include as date heading.}

\item{enc}{Text string to use as document encoding.}
}
\value{
An assembled html output page containing the required toc code.
}
\description{
Returns header structure for use with 'access_head()' and 'access_rmd()'.
}
