% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_alt_len.R
\name{check_alt_len}
\alias{check_alt_len}
\title{Check Rmd files for alt length.}
\usage{
check_alt_len(rmd_path = NULL, lan = detect_html_lang(lines))
}
\arguments{
\item{rmd_path}{Path to the Rmd that requires accessible header metadata. Rmd
must be output type html.}

\item{lan}{Identify the language of text content. Attempts to find a lang
attribute value from the rmd document. Alternatively, use a character string
such as "en".}
}
\value{
A line limit for alt text.
}
\description{
Return language specific alt text length limits. Limits are: eng	100,
ger	115, kor	90.
}
