% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compat.R
\name{check_compat}
\alias{check_compat}
\title{Checks if the html doctype is compatible.}
\usage{
check_compat(yaml_txt)
}
\arguments{
\item{yaml_txt}{Output of 'readLines()' separated from body text.}
}
\value{
Error condition if non-standard html found.
}
\description{
Checks to ensure file is not a special flavour of html output, including
xaringan, ioslides, flexdashboard or slidy. Output is an error if
html doc is a special case.
}
