% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_html_lang.R
\name{detect_html_lang}
\alias{detect_html_lang}
\title{Searches a file for valid methods of setting HTML lang attribute.}
\usage{
detect_html_lang(lines = NULL)
}
\arguments{
\item{lines}{The output of 'readLines()' or 'handle_rmd_path()'.}
}
\value{
The lang attribute value if found. Stops if no value found.
}
\description{
Takes the output of 'readLines()' and looks for lang attribute value,
stopping if no valid value is found. Typically used on the output of
'handle_rmd_path()'.
}
