% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_imgs.R
\name{find_all_imgs}
\alias{find_all_imgs}
\title{Find any markdown or HTML syntax images within read lines.}
\usage{
find_all_imgs(lines = NULL)
}
\arguments{
\item{lines}{Read lines, default is NULL & typically used with output of
handle_rmd_path.}
}
\value{
A named vector with line numbers of any found images and their
values.
}
\description{
Check any lines for images and return the line numbers and values.
}
