% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_alt_lim.R
\name{find_alt_lim}
\alias{find_alt_lim}
\title{Find alt length limit for specific lang values.}
\usage{
find_alt_lim(lines = NULL, lan = detect_html_lang(lines))
}
\arguments{
\item{lines}{Any Rmarkdown or HTML file lines. Typically the output of
'handle_rmd_path()'.}

\item{lan}{Identify the language of text content. Attempts to find a lang
attribute value from the rmd document. Alternatively, use a character string
such as "en".}
}
\value{
A line limit for alt text.
}
\description{
Return language specific alt text length limits. Limits are: eng	100,
ger	115, kor	90.
}
