% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_theme.R
\name{find_theme}
\alias{find_theme}
\title{Searches YAML head for YAML theme parameters, returning the theme if found.}
\usage{
find_theme(lines = NULL)
}
\arguments{
\item{lines}{The output of 'readLines()' or 'handle_rmd_path()'. Ensure
'check_compat()' has been used prior.}
}
\value{
The theme value as a character string.
}
\description{
Use on YAML head text only. 'check_compat()' should be used prior to the use
of 'find_theme()'. Looks for YAML theme parameter value. Returns the theme
if found of YAML syntax 'null' if not found.
}
