% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_rmd_path.R
\name{handle_rmd_path}
\alias{handle_rmd_path}
\title{Check rmd path & file type.}
\usage{
handle_rmd_path(rmd_path = NULL)
}
\arguments{
\item{rmd_path}{Path to the Rmd to check.}
}
\value{
Reads rmd lines on success. Error if file does not exist or is not an
rmd file.
}
\description{
Checks rmd_path exists and that the file suffix is correct.
}
