% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_toc.R
\name{insert_toc}
\alias{insert_toc}
\title{Insert the required code for the specified toc type.}
\usage{
insert_toc(toc, header, text, lan)
}
\arguments{
\item{toc}{TOC status, FALSE, TRUE or "float".}

\item{header}{Metadata items wrapped with 'tags$header()'.}

\item{text}{Raw text required for Rmarkdown body.}

\item{lan}{lang attribute value.}
}
\value{
An assembled html output page containing the required toc code.
}
\description{
Inserts 'render_toc()' code chunk if toc is TRUE. Inserts YAML header if toc
is "float".
}
