% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_yaml.R
\name{insert_yaml}
\alias{insert_yaml}
\title{Insert the required code for the specified toc type.}
\usage{
insert_yaml(toc, header, text, lan, theme = "default", highlight = "null")
}
\arguments{
\item{toc}{TOC status, FALSE (the default) or TRUE.}

\item{header}{Metadata items wrapped with 'tags$header()'.}

\item{text}{Raw text required for Rmarkdown body.}

\item{lan}{lang attribute value.}

\item{theme}{Text styling to apply. Current valid value is "default" only.
All other themes present accessibility errors on testing.}

\item{highlight}{Currently only "null" is a valid, due to accessibility
errors found in all built-in highlight options.}
}
\value{
An assembled html output page containing the required toc code.
}
\description{
Inserts YAML header for floating toc if toc is TRUE.
}
