% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langs.R
\docType{data}
\name{langs}
\alias{langs}
\title{Language subtag codes}
\format{
An object of class \code{character} of length 9079.
}
\source{
\href{https://www.iana.org/assignments/language-subtag-registry/language-subtag-registry}{IANA subtag registry}
}
\usage{
data(langs)
}
\description{
Subtag codes taken from the iana subtag registry.
}
\keyword{datasets}
