% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_rmds.R
\name{retrieve_rmds}
\alias{retrieve_rmds}
\title{Find Rmd files within a provided file structure.}
\usage{
retrieve_rmds(search_dir = ".", recurse = TRUE, to_txt = FALSE)
}
\arguments{
\item{search_dir}{The directory to search. Defaults to ".".}

\item{recurse}{Defaults to TRUE. Should child directories be searched or
not.}

\item{to_txt}{Defaults to FALSE. If TRUE, writes found rmd paths in a txt to
the \code{search_dir}.}
}
\value{
Relative paths to all found Rmd files.
}
\description{
Recursively searches file structure for Rmarkdown files. Returns the found
Rmd relative paths.
}
