% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_heading.R
\name{return_heading}
\alias{return_heading}
\title{Return a HTML heading tag of the required level.}
\usage{
return_heading(txt, lvl, class)
}
\arguments{
\item{txt}{The text to be used as the heading text.}

\item{lvl}{A number to be used for the heading level. 1 == tags$h1() and so
on.}

\item{class}{A character string to use as the first class attribute. Do not
include "toc-ignore", this will be added.}
}
\value{
null if txt is length 0 or required heading level with class attr
}
\description{
Returns a HTML heading tag of any of the available levels (h1, h2, h3 and so
on). The class selectors applied will always include "toc-ignore", but an
additional selector may be included within the class parameter. If an object
of length zero is passed to the txt parameter, NULL is returned.
}
