% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_urls.R
\docType{data}
\name{rmd_path}
\alias{rmd_path}
\title{Check an Rmarkdown for broken links.}
\format{
An object of class \code{character} of length 1.
}
\usage{
rmd_path
}
\arguments{
\item{rmd_path}{Path to the Rmd that requires links to be checked. Rmd must
be output type html.}
}
\value{
Lines of any urls that respond with an error.
}
\description{
Check links within an Rmarkdown document for any urls that responds with an
error.
}
\keyword{datasets}
