% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\docType{data}
\name{tags}
\alias{tags}
\title{"tags" object from "htmltools".
License: GPL (>= 2)
\href{https://github.com/rstudio/htmltools}{htmltools GitHub}
\href{https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html}{gpl 2 terms}}
\format{
An object of class \code{list} of length 181.
}
\usage{
tags
}
\value{
The html "tags" from "htmltools".
}
\description{
Does not appear to be exported from htmltools namespace. Dependency has been
established with "utils::getFromNamespace"
}
\keyword{datasets}
