% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_img.R
\name{access_img}
\alias{access_img}
\title{Produce an accessible image or chart.}
\usage{
access_img(img = last_plot(), alt = NULL, wid = 500, ht = 500, dpi = 300)
}
\arguments{
\item{img}{The image to include as accessible HTML. Defaults to
'ggplot2::last_plot()'. Can be replaced with an image written to disc.}

\item{alt}{A character string describing the image for screen reader
accessibility.}

\item{wid}{Width of the image in pixels. Defaults to 500.}

\item{ht}{Height of the image in pixels. Defaults to 500.}

\item{dpi}{Resolution. Please see \code{?ggplot2::ggsave()} for details.}
}
\value{
Inline HTML with the necessary structure for screen reader
accessibility.
}
\description{
Reads in an image and produces the HTML structure expected by web
accessibility checkers such as WAVE. Also works as a wrapper or around
charts.
}
