% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_rmd.R
\name{access_rmd}
\alias{access_rmd}
\title{Produce an accessible Rmarkdown file template.}
\usage{
access_rmd(
  filenm = NULL,
  title = NULL,
  author = NULL,
  date = NULL,
  lan = NULL,
  subtitle = NULL,
  toc = FALSE
)
}
\arguments{
\item{filenm}{Required - The name of the file.}

\item{title}{Required - The document title.}

\item{author}{Required - The document author.}

\item{date}{Required - The author date.}

\item{lan}{Required - The HTML language to set as the value for the lang
attribute.}

\item{subtitle}{Optional - The document subtitle}

\item{toc}{Optional, defaults to FALSE. Should a table of contents be
included.}
}
\value{
An Rmarkdown file with an HTML head, populated with metadata
specified within the function parameters.
}
\description{
Creates an Rmarkdown template with the HTML structure required by screen
readers.
}
