% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_alt_len.R
\name{check_alt_len}
\alias{check_alt_len}
\title{Check Rmd files for alt length.}
\usage{
check_alt_len(rmd_path = NULL, lan = NULL)
}
\arguments{
\item{rmd_path}{Path to the Rmd that requires accessible header metadata. Rmd
must be output type html.}

\item{lan}{Identify the language of text content.}
}
\value{
Adjust the Rmd YAML provided to \code{rmd_path}, improving its
accessibility for screen readers. Only works with html output.
}
\description{
Check any Rmd images for language specific alt text length limits. Limits
used are: eng	100, ger	115, kor	90.
}
