% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_theme.R
\name{find_theme}
\alias{find_theme}
\title{Searches YAML head for YAML theme parameters, returning the theme if found.}
\usage{
find_theme(yaml = NULL)
}
\arguments{
\item{yaml}{A YAML header text string.}
}
\value{
The theme value as a character string.
}
\description{
Use on YAML head text only. 'check_compat()' should be used prior to the use
of 'find_theme()'. Looks for YAML theme parameter value. Returns the theme
if found or 'default' if no theme specified.
}
