% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_alt.R
\name{sus_alt}
\alias{sus_alt}
\title{Check an image for suspicious alt text.}
\usage{
sus_alt(rmd_path = NULL)
}
\arguments{
\item{rmd_path}{Path to the Rmd that contains image tags to check.}
}
\value{
Line numbers of images that has alt text equal to placeholder values.
}
\description{
Checks if an image's alt text is equal to alt attribute placeholder values,
including: 'nbsp', 'spacer' and src attribute value (filename).
}
