% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tags}
\alias{tags}
\title{"tags" object from "htmltools".
License: GPL (>= 2)
\href{https://github.com/rstudio/htmltools}{htmltools GitHub}}
\value{
Wraps provided paramaters in HTML tag specified.
}
\description{
"tags" object from "htmltools".
License: GPL (>= 2)
\href{https://github.com/rstudio/htmltools}{htmltools GitHub}
}
\keyword{internal}
